fix: longValue() changed for doubleValue()

+++ binance4j-strategy/src/main/java/com/binance4j/strategy/dto/SymbolBar.java
@@ -27,11 +27,11 @@ public class SymbolBar extends BaseBar {
 	public SymbolBar(Bar bar, String symbol) {
 		super(bar.getTimePeriod(),
 				bar.getEndTime(),
-				bar.getOpenPrice().longValue(),
-				bar.getHighPrice().longValue(),
-				bar.getLowPrice().longValue(),
-				bar.getClosePrice().longValue(),
-				bar.getVolume().longValue());
+				bar.getOpenPrice().doubleValue(),
+				bar.getHighPrice().doubleValue(),
+				bar.getLowPrice().doubleValue(),
+				bar.getClosePrice().doubleValue(),
+				bar.getVolume().doubleValue());
 		this.symbol = symbol;
 	}
 }
